requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.7
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.5.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1 